requirements.txt
numpy>=1.24.4,<3.0
Pillow>=10.0
PyYAML>=6.0
matplotlib>=3.7
tqdm>=4.66
scikit-learn>=1.3
rasterio>=1.3
faiss-cpu>=1.7.4
pandas>=2.0
tifffile>=2023.9.26
torch
torchvision
